% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Create SUS-SIA/SIH database}
\usage{
create_output(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  health_establishment_id = "all",
  chunk_size
)
}
\arguments{
\item{year_start}{Ano de início da realização do procedimento}

\item{month_start}{Mês de início da realização do procedimento}

\item{year_end}{Ano de término da realização do procedimento}

\item{month_end}{Mês de término da realização do procedimento}

\item{state_abbr}{Sigla da Unidade Federativa}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\description{
\link{create_output} preprocess microdata files from SIA and SIH information systems (DATASUS)
and match with CNES and SIGTAP information.
}
\examples{
\dontrun{create_output(year_start=2021,
                  month_start=1,
                  year_end=2021,
                  month_end=3,
                  state_abbr="CE",
                  health_establishment_id=c("2561492","2481286")
                  )}
}
