% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output_from_local}
\alias{create_output_from_local}
\title{Create SUS-SIA/SIH database from local DBC files}
\usage{
create_output_from_local(dbc_dir_path, health_establishment_id)
}
\arguments{
\item{dbc_dir_path}{Diretório que contêm os arquivos DBC}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\description{
\link{create_output_from_local} preprocess DBC files from SIA and SIH information systems (DATASUS)
and match with CNES and SIGTAP information.
}
\examples{
\dontrun{create_output(tempdir(),
                       health_establishment_id=c("2561492","2481286")
                       )}
}
