% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_SIA.R
\name{create_output_SIA}
\alias{create_output_SIA}
\title{Create SUS-SIA database}
\usage{
create_output_SIA(year_start, month_start, year_end, month_end, uf, CNES)
}
\arguments{
\item{year_start}{Ano de início da realização do procedimento}

\item{month_start}{Mês de início da realização do procedimento}

\item{year_end}{Ano de término da realização do procedimento}

\item{month_end}{Mês de término da realização do procedimento}

\item{uf}{Sigla da Unidade Federativa}

\item{CNES}{Código(s) do estabelecimento de saúde}
}
\description{
\link{create_output_SIA} preprocess microdata files from SIA-PA information system (DataSUS)
and match with CNES and SIGTAP information.
}
\examples{
create_output_SIA(year_start=2021,
                  month_start=1,
                  year_end=2021,
                  month_end=3,
                  uf="CE",
                  CNES=c("2561492","2481286")
                  )
}
