% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_SIA.R
\name{get_counties_by_state}
\alias{get_counties_by_state}
\title{Create SUS-SIA database}
\usage{
get_counties_by_state(state_abbr)
}
\description{
\link{create_output_SIA} preprocess microdata files from SIA-PA information system (DataSUS)
and match with CNES and SIGTAP information.
}
