% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faturamento.sus-package.R
\docType{package}
\name{faturamento.sus-package}
\alias{faturamento.sus}
\alias{faturamento.sus-package}
\title{faturamento.sus: Preprocess SIA/SIH DATASUS database}
\description{
Preprocess SIA/SIH DATASUS and match with CNES and SIGTAP informations
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tivorn/faturamento.sus}
  \item Report bugs at \url{https://github.com/tivorn/faturamento.sus/issues}
}

}
\author{
\strong{Maintainer}: Victor Fernandes \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
