% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faturamento.sus-package.R
\docType{package}
\name{faturamento.sus-package}
\alias{faturamento.sus}
\alias{faturamento.sus-package}
\title{faturamento.sus: Preprocess SIA-PA DATASUS database}
\description{
Preprocess SIA-PA DATASUS and match with CNES and SIGTAP informations
}
\author{
\strong{Maintainer}: Victor Fernandes \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
